+++ src/client/runtime/WebpackHMRClient.ts
@@ -1,7 +1,7 @@
 /* eslint-env browser */
 /* globals __webpack_hash__ __DEV__ */
 
-import type { HMRMessage, HMRMessageBody } from '../types';
+import type { HMRMessage, HMRMessageBody } from '../../types';
 import { getDevServerLocation } from './getDevServerLocation';
 
 class HMRClient {
